Add unit tests through unity test runner
As per the unity guide, this requires creating an assembly that
depends on the `nunit.framework.dll` reference. In order to
allow this testing assembly to access the code under the OpenTS2
folder, we create a `ScriptsAssembly` assembly that the `Tests`
assembly can depend on.

Unfortunately these tests can only be run through unity or the
Rider IDE but unless we want to do a clean separation of all the
Unity bits from the C# code this will likely be the easiest
option.