Add loading of GMDC static/bone bounding meshes
This only loads the last 2 sections of a GMDC file, where the
bounding meshes are stored as these are the simplest to import.
This lays out the basic framework for GMDC loading and adds a
little test file in the form of the utah teapot so we have a
copyright free asset to test against.