Add gameobject to transform from sims2 space to unity
Previously we were using an ad-hoc system where we were rotating
meshes to -90 on the x-axis and applying a -1 scale to them. This
worked to show them properly but meant that because our coordinate
system was flipped all rotations from the game had to be inverted
from a right-hand system to a left-hand system.

This change introduces a simRotations object that all the meshes
sit under that takes raw rotations from the game. The parent game
object then converts these all at once into the unity space.

All rotations from the game should now be applied to this simRotations
object, we should likely move this ScenegraphResourceAsset generated
game object into a helper MonoBehavior to make this easier to use.